+++ js/page-projects.js
@@ -50,9 +50,6 @@ for (let i = projects.length - 1; i >= 0; i--) {
     projectTagLang.className = "project-tag";
     projectTagLang.textContent = project.languages[0].name;
     projectTagLang.style.backgroundColor = project.languages[0].color;
-    if (project.languages[0].whiteText) {
-        projectTagLang.style.color = "#C0C0C0";
-    }
     projectTags.push(projectTagLang);
 
     var projectTagLic = document.createElement("p");
@@ -105,9 +102,6 @@ for (let i = projects.length - 1; i >= 0; i--) {
         languageBar.textContent = language.name.replace("JavaScript", "Java\u00ADScript");
         languageBar.style.width = width.toString() + "px";
         languageBar.style.backgroundColor = language.color;
-        if (language.whiteText) {
-            languageBar.style.color = "#C0C0C0";
-        }
         if (width < 75) {
             languageBar.style.direction = "rtl";
         }